web/requirements.txt
flask==2.1.2
sqlalchemy==1.4.36
requests==2.27.1
psycopg2-binary==2.9.3
gunicorn==20.1.0
